chore:moved the mock data to the mock-data file and removed the import of built component
